fix ui_subghzd.cpp:85:29: warning: narrowing conversion of 'this->ui::SubGhzDRecentEntryDetailView::cnt' from 'uint32_t' {aka 'long unsigned int'} to 'uint16_t' {aka 'short unsigned int'} [-Wnarrowing]

+++ firmware/application/keeloq_file.hpp
@@ -30,7 +30,7 @@
 struct KeeloqData {
     std::string mf_name{};
     uint32_t serial = 0;
-    uint16_t counter = 0;
+    uint32_t counter = 0;
     uint8_t btn = 0;
 };
